SOC-5015: Eliminate kmallocs in SPU driver
Eliminate kmallocs in the SPU driver request path to
improve performance. Embed the msg buf structure within
the rctx structure, which is allocated by the caller for
each request. Callers may not zero the memory. So memset
certain buffers to 0. For hash export, do not include the
msg buffers in the state size and do not include the buffers
when exporting or importing the hash state.

Bug: SOC-5015
Change-Id: Iac718b8957a476cceea594f7fcc4af15b6dc8b4f
Signed-off-by: Raveendra Padasalagi <raveendra.padasalagi@broadcom.com>
Signed-off-by: Rob Rice <Rob.Rice@broadcom.com>
Reviewed-on: http://gerrit-ccxsw.rtp.broadcom.com:8080/35735
Reviewed-by: CCXSW <ccxswbuild@broadcom.com>
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Tested-by: CCXSW <ccxswbuild@broadcom.com>
Tested-by: vpx_smoketest status <vpx_smoketest@broadcom.com>
Reviewed-by: Scott Branden <scott.branden@broadcom.com>