VOIP-6743:  Create ASoC dummy driver for BT voice link
Create a dummy codec driver for use with the Cypress BT
chips that we use on any of our reference designs.
This driver more closely models that behevior of the Cypress
BT voice link than the generic BT voice driver.  For example,
our driver can enforce the constant bits per second rule
that is needed in tdm mode.

Bug: VOIP-6743
Change-Id: I96322a317a4ec1e7324e2fc031a03d659d4d2639
Reviewed-on: http://gerrit-ccxsw.broadcom.net/35968
Reviewed-by: Ray Jui <ray.jui@broadcom.com>
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Tested-by: vpx_smoketest status <vpx_smoketest@broadcom.com>
Tested-by: Lori Hikichi <lori.hikichi@broadcom.com>
Reviewed-by: Scott Branden <scott.branden@broadcom.com>