VOIP-6743: Create ASoC dummy driver for BT voice link
Create a dummy codec driver for use with the Cypress BT
chips that we use on our reference designs.
This driver more closely models that behavior of the Cypress
BT voice link than the generic BT voice driver.  For example,
our driver can enforce the constant bits per second rule
that is needed in tdm mode.

Bug: VOIP-6743
Change-Id: I0995c374ddaebc890a5a2ede9ffc214acee8c293
Reviewed-on: http://gerrit-ccxsw.broadcom.net/36185
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Reviewed-by: Ray Jui <ray.jui@broadcom.com>
Tested-by: vpx_smoketest status <vpx_smoketest@broadcom.com>
Reviewed-by: CCXSW <ccxswbuild@broadcom.com>
Tested-by: CCXSW <ccxswbuild@broadcom.com>
Tested-by: Lori Hikichi <lori.hikichi@broadcom.com>
Reviewed-by: Scott Branden <scott.branden@broadcom.com>