VOIP-6750:  Cyngus IP phone device tree changes to enable sound
Add device tree changes for the bcm11360 enterprise IP Phone
to enable the ASoC machine driver for the bcm911360 (Cygnus) IP phone
reference design.
The platform consists of a BCM911360 SoC and CS42L73 codec.  There is also
the option have having a Bluetooth voice link on a second DAI.

Bug: VOIP-6750
Change-Id: I35b49de940a8a35ff5b2ac47e8e3a781a7c34e68
Reviewed-on: http://gerrit-ccxsw.broadcom.net/36459
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Tested-by: vpx_smoketest status <vpx_smoketest@broadcom.com>
Reviewed-by: CCXSW <ccxswbuild@broadcom.com>
Tested-by: CCXSW <ccxswbuild@broadcom.com>
Reviewed-by: Ray Jui <ray.jui@broadcom.com>
Reviewed-by: Scott Branden <scott.branden@broadcom.com>