VOIP-6760:  Fix ASoC DAPM errors in CS42L73 codec driver
The CS42L73 is reporting problems during driver initialization.
There are several errors, but they all have the same form.

> Control not supported for path  XXXX -> YYYY -> ZZZZ
> ASoC: no dapm match for XXXX --> YYYY --> ZZZZ
> ASoC: Failed to add route XXXX -> YYYY -> ZZZZ

The driver has never changed regarding this aspect, so the problems
are likely due to change in the ASoC framework in a recent kernel
version.  Setting the mixer control to NULL in the DAPM entry fixes
the problem. The mixer control is not needed, so setting it to NULL
make sense.

More info from Lori:
The codec driver itself has always been broken.
In the past the framework did not flag this as a problem,
but now there is some additional error checking in the framework.
It seems the error is rather benign because it never used to cause
 us a problem. Setting this entry to NULL does disable anything
because that aspect of the dapm route never worked properly anyway.
The driver works just as well (or poorly) as it used to in previous
kernel versions.

The reason this might be considered a workaround is because the
original designer was probably intending to accomplish something here,
I just do not know what. OR perhaps it was an outright mistake that
they filled in this field. Who knows? There is never any comments added
to drivers. Until I go through an analysis of the codecs block diagram
and map it into what has been coded I can not say for sure.

Change-Id: I0003889069a418cd4f7219bd26c9281a9f6d9357
Reviewed-on: http://gerrit-ccxsw.broadcom.net/36618
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Tested-by: vpx_smoketest status <vpx_smoketest@broadcom.com>
Reviewed-by: Ray Jui <ray.jui@broadcom.com>
Reviewed-by: Scott Branden <scott.branden@broadcom.com>