VOIP-6760: Fix error in cs42l73 driver when setting mclk clk id
The cs42l73 codec driver is not setting the mclksel field prior
to calling a function that uses the value.  The value defaults
to CS42L73_CLKID_MCLK1, therefore, the problem would only be
noticed we were trying to use CS42L73_CLKID_MCLK2.
The driver does set the mclksel field later in the sequence,
which then results in everything working properly if we attempt
the whole operation again using CS42L73_CLKID_MCLK2.

Bug: VOIP-6760
Change-Id: I854643f2c52acaf8a5b18663ed0c52e08512ebd9
Reviewed-on: http://gerrit-ccxsw.broadcom.net/37054
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Reviewed-by: CCXSW <ccxswbuild@broadcom.com>
Reviewed-by: Ray Jui <ray.jui@broadcom.com>
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Tested-by: CCXSW <ccxswbuild@broadcom.com>
Tested-by: vpx_smoketest status <vpx_smoketest@broadcom.com>
Reviewed-by: Scott Branden <scott.branden@broadcom.com>