SOC-5153: Add PAXC PHY support
phy: sr-pcie: add Stingray PAXC PHY support

This patch adds a dummy PAXC PHY support to the Stingray PCIe PHY
driver. This is done by querying the MHB/PAXC power status in the PHY
init callback. In the case when MHB/PAXC is unpowered, the PHY driver
will fail to initialize and as a result it will prevent the PCIe host
driver from proceeding to enumerate with Nitro

Signed-off-by: Ray Jui <ray.jui@broadcom.com>

Bug: SOC-5153
Change-Id: I42ebd90ca20c22bc5b6a0d4377bf090d0650d4d2
Reviewed-on: http://gerrit-ccxsw.broadcom.net/37060
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Reviewed-by: CCXSW <ccxswbuild@broadcom.com>
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Tested-by: vpx_smoketest status <vpx_smoketest@broadcom.com>
Tested-by: CCXSW <ccxswbuild@broadcom.com>
Reviewed-by: Scott Branden <scott.branden@broadcom.com>