VOIP-6760: Fix error in cs42l73 DAPM entries for mic power down
The DAPM entries used to control the power down of the digital
mics and ADC are labelled backwards regarding left and right.
The datasheet lists these register enteries as A and B, without
any orientation of left and right, but the convention throughout
the datasheet is that A is left and B is right channel.

Bug: VOIP-6760
Change-Id: I293fe94d6d0375942d2726920c371d2e0ddc1fb2
Reviewed-on: http://gerrit-ccxsw.broadcom.net/37223
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Tested-by: vpx_smoketest status <vpx_smoketest@broadcom.com>
Reviewed-by: Ray Jui <ray.jui@broadcom.com>
Reviewed-by: Scott Branden <scott.branden@broadcom.com>