VOIP-6819: Remove SET_PARENT property from audio pll post divider clocks
Functionality was added within the audio clocking tree to block a parent
clock change if there are multiple child clocks active. While this make
sense, it is confusing for the end user who may not be aware of another
channel's state.
Changing the parent was initially added to achieve the goal of producing
PLL Vco frequencies that were ideal for both 44.1 kHz and 48 kHz frame
rates.  A cleaner approach is to use a single PLL Vco value that works
for both domains. This has the downside of limiting the range of
frequencies in each domain, but there should still be enough available
to satisfy our use cases.
Therefore, in addition to removing the SET PARENT property, several
other predefined Vco frequencies have been added to the configuration
table to give the system designer a range of choices.
One other note, the calculation in iproc_clk_determine_rate() was
incorrect.  DIV_ROUND_UP has been replaced with DIV_ROUND_CLOSEST
because we want to find the divisor that gives us the smallest delta.

Bug: VOIP-6819
Change-Id: I40d3377dae2a7bb5bf020024cfcca5f123de3d6b
Reviewed-on: http://gerrit-ccxsw.broadcom.net/37169
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Tested-by: vpx_smoketest status <vpx_smoketest@broadcom.com>
Reviewed-by: Ray Jui <ray.jui@broadcom.com>
Reviewed-by: Scott Branden <scott.branden@broadcom.com>