VOIP-6762: Cygnus audio allow support for 384 kHz
The ALSA audio framework does not offer support for 384 kHz
as one of its commmon known rates, i.e, it is not enumerated.
The rate can be supported through the use of SNDRV_PCM_RATE_KNOT.
In addition to allowing 384 kHz, it makes sense in this commit
to revise the method used to validate and calculate the necessary
register settings.  The old method use a lookup table. The new
method uses a ratio rule and known maximum and minimum values.

Bug: VOIP-6762
Change-Id: Ib98c99aa9b8abf0d5c71c91ceb5d2a166080a89e
Reviewed-on: http://gerrit-ccxsw.broadcom.net/37336
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Tested-by: vpx_smoketest status <vpx_smoketest@broadcom.com>
Reviewed-by: CCXSW <ccxswbuild@broadcom.com>
Tested-by: CCXSW <ccxswbuild@broadcom.com>
Reviewed-by: Ray Jui <ray.jui@broadcom.com>
Reviewed-by: Scott Branden <scott.branden@broadcom.com>