SOC-5099: eth Interface doesn't process few packets
this patch fixes eth interface dropping packets because
of skb alloc error, this happens partly because poll weight
is 64, while internl sram size is limited. so when IXIA
sends packets, Transmit tasklet delays freeing the Tx memory.
also poll weight should be of a reasonable value in order to have
proper flow control. so we have to bring down NAPI quota.

Also it takes care of solving IXIA sending more packets and
packets are being dropped.

but since external SRMA is slow and there is
double copy involved, which inserts lot of wait cycles
fetching packets. so this test is supposed to fail on
external SRAM with 1G. but the same passes with 10M interface,
of course again it can not handle huge traffic.

also it implements netif_stop_queue in case of transmit fifo
is full, to ensure flow control between network stack and device queue.

Bug: SOC-5099
Change-Id: Ifbc8a93114309fa7c1f5d68799db4cf5cd6786c2
Signed-off-by: Oza Pawandeep <oza@broadcom.com>
Reviewed-on: http://gerrit-ccxsw.broadcom.net/37348
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Reviewed-by: CCXSW <ccxswbuild@broadcom.com>
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Tested-by: vpx_smoketest status <vpx_smoketest@broadcom.com>
Tested-by: CCXSW <ccxswbuild@broadcom.com>
Reviewed-by: Ray Jui <ray.jui@broadcom.com>
Reviewed-by: Arun Parameswaran <arun.parameswaran@broadcom.com>
Reviewed-by: Scott Branden <scott.branden@broadcom.com>