VOIP-6819: Cyngus pll use determine_rate instead for round_rate
Move the Cygnus PLL common clock code to use determine_rate api
instead of round_rate.  The round_rate api suffers from a problem
when the returned rounded rate is larger than about 2.1 GHz.
The return type is signed so the value crosses the 31 bit
threshold.  The returned rate will look like and error code to
the clocking framework instead of a valid rate.  determine_rate
does not suffer from this problem and is the intended replacement
for round_rate.

Bug: VOIP-6819
Change-Id: I24f9118e3e8f028a21cb72bead26b570021522f0
Reviewed-on: http://gerrit-ccxsw.broadcom.net/37394
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Reviewed-by: CCXSW <ccxswbuild@broadcom.com>
Reviewed-by: Ray Jui <ray.jui@broadcom.com>
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Tested-by: vpx_smoketest status <vpx_smoketest@broadcom.com>
Tested-by: CCXSW <ccxswbuild@broadcom.com>
Reviewed-by: Scott Branden <scott.branden@broadcom.com>