VOIP-6762: Modify how Cygnus audio manages incoming clocks
Each audio port is now assigned a clock via the devicetree.
The previous code had a problem when mapping the clocks to
port so this commit also fixes that issue.  Also, much of
the suspend and resume logic in the audio driver has been
removed because it was just dealing with the clocks.  This
is not needed because any needed supsend resume code should
be handled be the clock driver.

Bug: VOIP-6762
Change-Id: Ifd64478f383d1bb64b85bb5a3f0024ff32e94130
Reviewed-on: http://gerrit-ccxsw.broadcom.net/37437
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Tested-by: vpx_smoketest status <vpx_smoketest@broadcom.com>
Reviewed-by: CCXSW <ccxswbuild@broadcom.com>
Tested-by: CCXSW <ccxswbuild@broadcom.com>
Reviewed-by: Ray Jui <ray.jui@broadcom.com>
Reviewed-by: Scott Branden <scott.branden@broadcom.com>