VOIP-6762: Cygnus audio add clock helper functions
Add helper functions to get and put a clock resource for an audio port.
Some codecs require that the SOC provide an MCLK early in the
initialization phase.  These helper functions provide a way for the
machine driver to get access to the clock device which is being
managed by the SOC audio driver.  The machine driver can then
initialize the clock early in the startup sequence.

Bug: VOIP-6762
Change-Id: I05ff5ada97534caf0317f79672d3c16e2938e9e2
Reviewed-on: http://gerrit-ccxsw.broadcom.net/37943
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Tested-by: vpx_smoketest status <vpx_smoketest@broadcom.com>
Tested-by: CCXSW <ccxswbuild@broadcom.com>
Reviewed-by: CCXSW <ccxswbuild@broadcom.com>
Reviewed-by: Ray Jui <ray.jui@broadcom.com>
Reviewed-by: Scott Branden <scott.branden@broadcom.com>