VOIP-5343: Add ASoC codec for BRCM generic test codec
This driver is very similar to the "dummy" driver found in the ASoC
framework, but we are allowing this driver to be continuous frame rate
up to a maximum of 384 kHz.  It is meant to model a piece of hardware
that may be attached to a port for such uses as testing and development
End products should create or make use of a more specific codec driver.

Bug: VOIP-5343
Change-Id: Ieb8723ea29e8a21fcdd8c4ebf1b0901f25ae9622
Reviewed-on: http://gerrit-ccxsw.broadcom.net/38267
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Tested-by: vpx_smoketest status <vpx_smoketest@broadcom.com>
Reviewed-by: CCXSW <ccxswbuild@broadcom.com>
Tested-by: CCXSW <ccxswbuild@broadcom.com>
Reviewed-by: Scott Branden <scott.branden@broadcom.com>