VOIP-6819: Fix calculation error in Cygnus' clk_set_rate()
The code was using DIV_ROUND_UP but should be using
DIV_ROUND_CLOSEST to find the divider with the the smallest
delta.  This problem would be most obvious if best divider
had a small (non-zero) remainder.  We often configure the
system to choose frequencies that are prefect multiples
of the various clocks so this problem was not seen for
most usage case.  The same problem was fixed in round_rate(),
but this instance of the calculation was missed.

Bug: VOIP-6819
Change-Id: I4eb83523f718759d934183a065ff7924c0699d56
Reviewed-on: http://gerrit-ccxsw.broadcom.net/38339
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Reviewed-by: CCXSW <ccxswbuild@broadcom.com>
Reviewed-by: Ray Jui <ray.jui@broadcom.com>
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Tested-by: vpx_smoketest status <vpx_smoketest@broadcom.com>
Tested-by: CCXSW <ccxswbuild@broadcom.com>
Reviewed-by: Scott Branden <scott.branden@broadcom.com>