SOC-5313: PCI: iproc: retry request when CRS returned from EP
PCI: iproc: Retry request when CRS returned from EP

For Configuration Requests only, following reset
it is possible for a device to terminate the request
but indicate that it is temporarily unable to process
the Request, but will be able to process the Request
in the future – in this case, the Configuration Request
Retry Status 10 (CRS) Completion Status is used

SPDK user space nvme driver unbind kernel driver followed by
nvme_remove, and following that it tries to bind use space driver
and while doing this some configuration requests get NAKed.
current iproc PCI driver is agnostic about it. PAXB will forward
the NAKed response in stipulated AXI code.

this patch fixes the problem, and attempts to read again in case
of PAXB forwarding the NAK. this is inline with PCI spec.
it implements iproc_pcie_config_read which gets called for Stingray.
otherwise it falls back to pci generic APIs.

Bug: SOC-5313
Change-Id: Ide8278e7586fce4a31f1d0dcff2f71c3ace15787
Signed-off-by: Oza Pawandeep <oza.oza@broadcom.com>
Reviewed-on: http://gerrit-ccxsw.broadcom.net/39586
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Reviewed-by: Ray Jui <ray.jui@broadcom.com>
Reviewed-by: CCXSW <ccxswbuild@broadcom.com>
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Tested-by: vpx_smoketest status <vpx_smoketest@broadcom.com>
Tested-by: CCXSW <ccxswbuild@broadcom.com>
Reviewed-by: Scott Branden <scott.branden@broadcom.com>