VOIP-6888: misc: dte: Add support for poll interface
The patch adds support for poll() interface in the DTE driver.
This enables the user-space apps to use poll() or select()
APIs to wait for timestamps to be available from the clients
and then read the timestamp.

The patch also enforces 'one client per file descriptor' rule.
File pointers are associated to the client id, that they enable
via ioctls. This association is used in the poll interface to
register wait queues for the client.

Bug: VOIP-6888
Change-Id: I38ff0b187398efda0b4ba41e64e163acfc6c425f
Reviewed-on: http://gerrit-ccxsw.broadcom.net/40235
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Reviewed-by: Arun Parameswaran <arun.parameswaran@broadcom.com>
Tested-by: Arun Parameswaran <arun.parameswaran@broadcom.com>
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Reviewed-by: Ray Jui <ray.jui@broadcom.com>
Tested-by: CCXSW <ccxswbuild@broadcom.com>
Reviewed-by: CCXSW <ccxswbuild@broadcom.com>
Reviewed-by: Scott Branden <scott.branden@broadcom.com>