VOIP-6903: Allow iproc plls to calculate vco params
The current iproc pll code uses lookup tables to find the pll parameters
for a limited set of frequencies.  Add the ability for the plls to
calculated their needed parameters.  This allows the user to select from
the full range of pll possible frequencies.
The old method is still used be default and the newer calculation is
being enabled for the audiopll.

Bug: VOIP-6903
Change-Id: Id4cb549a7bc950f28d774289717270dc1325b7d8
Reviewed-on: http://gerrit-ccxsw.broadcom.net/40339
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Reviewed-by: CCXSW <ccxswbuild@broadcom.com>
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Tested-by: vpx_smoketest status <vpx_smoketest@broadcom.com>
Tested-by: CCXSW <ccxswbuild@broadcom.com>
Reviewed-by: Ray Jui <ray.jui@broadcom.com>
Reviewed-by: Scott Branden <scott.branden@broadcom.com>