SOC-4653: Use changed stream-ids and msi-device-ids for FS4
arm64: dts: Use changed stream-ids and msi-device-ids for FS4

The FS4 stream-ids and msi-device-ids use same numbering space
because both are passed via same AXI4 user signals. This also
means that for MSI writes stream-id is same as msi-device-id.

Based on above, we have changed stream-ids and msi-device-ids
assigned to FS4 engines in ATF such that FS4 msi-device-ids do
not overlap with stream-id space of other blocks.

This patch updates stingray DTS to be in-sync with ATF.

Signed-off-by: Anup Patel <anup.patel@broadcom.com>
Bug: SOC-4653
Change-Id: I514cd81a985c282862b25155b1be11fc772491e1
Reviewed-on: http://gerrit-ccxsw.broadcom.net/40364
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Reviewed-by: CCXSW <ccxswbuild@broadcom.com>
Tested-by: CCXSW <ccxswbuild@broadcom.com>
Reviewed-by: Scott Branden <scott.branden@broadcom.com>