VOIP-6878: Add support for building Cygnus in secure mode.
arm: mach-bcm: This change allows the kernel to build the
bcm_cygnus for secure mode boot.

The CONFIG_OPTEE flag will be enabled only for the platforms, that
require the Kernel to boot in non-secure mode.

The defconfig file associated with the secure boot, should not
set the CONFIG_OPTEE flag for the secure boot to work.

It should also be noted that the secure boot does not support the
TrustZone support and so there will be no secure services in the
secure boot.

Bug: VOIP-6878
Change-Id: I9b518c746bc099ca3f77da4942a2a4d8adaf4232
Tested-by: ismail <ismail@broadcom.com>
Signed-off-by: ismail <ismail@broadcom.com>
Reviewed-on: http://gerrit-ccxsw.broadcom.net/40395
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Reviewed-by: Jonathan Richardson <jonathan.richardson@broadcom.com>
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Tested-by: vpx_smoketest status <vpx_smoketest@broadcom.com>
Reviewed-by: Shamez Kurji <shamez.kurji@broadcom.com>
Reviewed-by: CCXSW <ccxswbuild@broadcom.com>
Tested-by: CCXSW <ccxswbuild@broadcom.com>
Reviewed-by: Scott Branden <scott.branden@broadcom.com>