SOC-4556 Allow BCM5481x PHYs to setup internal TX/RX clock delay as required
This patch allows BCM5481x series PHYs to provide or disable internal
TX and/or RX clock delay so as to satisfy RGMII timing specifications
on a given platform.

On a particular platform, whether TX and/or RX clock delay is required
depends on how PHY connected to the MAC IP. This requirement can be
specified through "phy-mode" property in the platform device tree.

Change-Id: I156a251af13f4bf1c7cdd969a3389e6190aa650c
Signed-off-by: Abhishek Shah <abhishek.shah@broadcom.com>
Signed-off-by: Scott Branden <scott.branden@broadcom.com>
Reviewed-on: http://gerrit-ccxsw.broadcom.net/40593
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Reviewed-by: Ray Jui <ray.jui@broadcom.com>
Reviewed-by: CCXSW <ccxswbuild@broadcom.com>
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Tested-by: vpx_smoketest status <vpx_smoketest@broadcom.com>
Tested-by: CCXSW <ccxswbuild@broadcom.com>