SOC-5397: spu2: add support for single SPU2 dt
Existing SPU2 driver is designed assuming each SPU engine
will have only one mailbox channel to access it and hence in
dts file user needed to provide number of SPU dt nodes as many
as we have in the SoC.

Stingray has 10 SPU-engines and these engines are accessed/
programmed through 32 Flex-RM rings. Each Flex-RM ring is mapped
to one mailbox channel. So with the existing driver we can't support
stingray.

So this patch enhances driver to accept single SPU2 DT node and mention
all the mailbox instances in that node.

Bug:SOC-5397
Change-Id: I664521634c3e70eb1b7b76d1de3cbdc69ffebb39
Signed-off-by: Raveendra Padasalagi <raveendra.padasalagi@broadcom.com>
Reviewed-on: http://gerrit-ccxsw.broadcom.net/40870
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Tested-by: vpx_smoketest status <vpx_smoketest@broadcom.com>
Reviewed-by: CCXSW <ccxswbuild@broadcom.com>
Tested-by: CCXSW <ccxswbuild@broadcom.com>
Reviewed-by: Ray Jui <ray.jui@broadcom.com>
Reviewed-by: Scott Branden <scott.branden@broadcom.com>