SOC-5216: nvme: Implement Interrupt Coalescing Support
nvme: Implement Interrupt Coalescing Support

this patch enables interrupt coalescing feature
of NVMe.

According to NVMe spec:
The controller should signal an interrupt when either
the Aggregation Time or the Aggregation Threshold
conditions are met. If either the Aggregation Time or
the Aggregation Threshold fields are cleared to 0h, then
an interrupt may be generated (i.e., interrupt coalescing
is implicitly disabled).

The controller may delay an interrupt if it detects that
interrupts are already being processed for this vector.
Specifically, if the Completion Queue Head Doorbell register
is being updated that is associated with a particular interrupt
vector, then the controller has a positive indication that
completion queue entries are already being processed.
In this case, the aggregation time and/or the aggregation
threshold may be reset/restarted upon the associated register
write. This may result in interrupts being delayed indefinitely
in certain workloads where the aggregation time or aggregation
threshold are non-zero.

this patch takes interrupt coalesce (timeout and threshold)
value as module param, which could vary depending on SOCs
particularly RC's capabilities. which can be passed at boot-time.
If not then, by default this feature stays disabled.

this improves the free cpu cycles/cpu usage by 20%.

Bug: SOC-5216
Change-Id: I885d49a663314b5538f05264754a6d3dcd6329da
Signed-off-by: Oza Pawandeep <oza.oza@broadcom.com>
Reviewed-on: http://gerrit-ccxsw.broadcom.net/40886
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Reviewed-by: Ray Jui <ray.jui@broadcom.com>
Reviewed-by: CCXSW <ccxswbuild@broadcom.com>
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Tested-by: vpx_smoketest status <vpx_smoketest@broadcom.com>
Tested-by: CCXSW <ccxswbuild@broadcom.com>
Reviewed-by: Scott Branden <scott.branden@broadcom.com>