VOIP-6762: Populate Cygnus' ASoC function for set_pll
Provide the function allowing the ASoC machine driver to control Cygnus'
audio pll.  This ability is needed for the situations when the PLL
frequency chosen via the devicetree setting needs to be changed.
Typical this situation will arise when the user wants to optimize the
system for a specific clock property that is not satisfied with a
single boot time chosen frequency (e.g., high precision for 48 Khz).

Bug: VOIP-6762
Change-Id: I1c60bd76dbf22e38ce8a1f290c68046c2956a0f1
Reviewed-on: http://gerrit-ccxsw.broadcom.net/41553
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Reviewed-by: Ray Jui <ray.jui@broadcom.com>
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Tested-by: vpx_smoketest status <vpx_smoketest@broadcom.com>
Reviewed-by: Scott Branden <scott.branden@broadcom.com>