SOC-4556 create new dtb's for FS4 HW acceleration support
Unforunately the FS4 acceleration drivers enable polling
hardware during the init stage of the driver.  This causes the FS4
hardware to be constantly polling the CCN/DDR.  This activity has
a negative side effect on system performance.  So, unless the fs4
acceleration drivers are changed to only enable the polling hardware on
open and turn it off on close the FS4 acceleration must be disabled
in common use cases.  Only when raid or crypto is needed in the system will
it be enabled in the dtb.  So we need to create a different dtb
for -fs4 enabled systems.  Hopefully this dtb can be removed if the
raid/crypto drivers can be changed to not enable polling on init.

Change-Id: I10fe1ebf8d6c984999d0f47b9c81537a90a208be
Reviewed-on: http://gerrit-ccxsw.broadcom.net/41802
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Reviewed-by: CCXSW <ccxswbuild@broadcom.com>
Reviewed-by: Velibor Markovski <Velibor.Markovski@broadcom.com>
Reviewed-by: Ray Jui <ray.jui@broadcom.com>
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Tested-by: vpx_smoketest status <vpx_smoketest@broadcom.com>
Tested-by: CCXSW <ccxswbuild@broadcom.com>
Reviewed-by: Scott Branden <scott.branden@broadcom.com>