SOC-4556 define HAVE_PCI_MMAP for arm64
define HAVE_PCI_MMAP for arm64 so sys entries are available.
Upstream is trying to phase this out somehow even thought it is
enabled on almost all other architectures.

And, we need if for user space DPDK to wor with kernel drivers.
Proper plumbing needs to be done with DPDK so this can be removed.

Change-Id: Ieb400b6bb10c9f34d3898669f4f22737f2ec6a3a
Reviewed-on: http://gerrit-ccxsw.broadcom.net/42013
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Reviewed-by: CCXSW <ccxswbuild@broadcom.com>
Reviewed-by: Scott Branden <scott.branden@broadcom.com>
Tested-by: Scott Branden <scott.branden@broadcom.com>