SOC-5377: PCI: iproc: reduce the timeout for CRS
PCI: iproc: reduce the timeout for CRS

The value of retry 0xffff0001 is valid BAR contents.
and one of the Samsung NVMe drive shows up that.

NVMe defines this timeout in 500 ms range, and this
only happens if controller has been in reset, with new firmware,
or in abrupt shutdown case. And meanwhile config access
could result into retry.
Although it is not recommended to send cfgs.
But anyway that has to be handled.

500 ms is enough for the timeout as far as SPDK is concerned.
this is ASIC bug which would be fixed in SR+, where they
would create separate register to indicate completion status

Bug: SOC-5377
Change-Id: Iede22ec9e27ec845f7c75484a35e5e8a430528e6
Signed-off-by: Oza Pawandeep <oza.oza@broadcom.com>
Reviewed-on: http://gerrit-ccxsw.broadcom.net/42182
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Reviewed-by: CCXSW <ccxswbuild@broadcom.com>
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Tested-by: vpx_smoketest status <vpx_smoketest@broadcom.com>
Tested-by: CCXSW <ccxswbuild@broadcom.com>
Reviewed-by: Ray Jui <ray.jui@broadcom.com>
Reviewed-by: Scott Branden <scott.branden@broadcom.com>