SOC-5401: Disable MSI parsing logic for PAXC in Stingray
PCI: iproc: Disable MSI parsing logic for PAXC in Stingray

Stingray has an ASIC bug in the PAXC block that does not process MSI
writes correctly. Therefore, the internal MSI processing logic needs to
be disabled. An external IOMMU block can be used to forward the MSI
writes to the GIC instead

This patch adds logic to disable this parsing logic for PAXC on
Stingray, based on matched device tree compatible strings

Change-Id: I6e493458d9709a80c70ae98826db5da46169ebb0
Signed-off-by: Ray Jui <ray.jui@broadcom.com>
Reviewed-on: http://gerrit-ccxsw.broadcom.net/42605
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Tested-by: vpx_smoketest status <vpx_smoketest@broadcom.com>
Tested-by: CCXSW <ccxswbuild@broadcom.com>
Reviewed-by: CCXSW <ccxswbuild@broadcom.com>
Reviewed-by: Scott Branden <scott.branden@broadcom.com>