VOIP-6956: Cygnus audio spdif does not need set_fmt
Remove the set set_fmt function from the SPDIF's DAI.
Previously we never noticed a problem with populating SPDIF's
set_fmt function because it was never called by any machine
drivers.  Now we are using a generic initialization routine
and setting the format statically.  This cause the ASoC
framework to call set_fmt, if it is populated.  This resulted
in an error for the SPDIF port.

Bug: VOIP-6956
Change-Id: I51b238fe4829cab0a85a81d56825fe35fcb60bdd
Reviewed-on: http://gerrit-ccxsw.broadcom.net/42942
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Reviewed-by: Ray Jui <ray.jui@broadcom.com>
Tested-by: vpx_smoketest status <vpx_smoketest@broadcom.com>
Reviewed-by: CCXSW <ccxswbuild@broadcom.com>
Tested-by: CCXSW <ccxswbuild@broadcom.com>
Reviewed-by: Scott Branden <scott.branden@broadcom.com>