SOC-5397: spu2: Remove BCM_PDC_MBOX dependency
This patch removes PDC_MBOX "depends on" for SPU driver
in Kconfig and adds MAILBOX as dependent module.
Selecting PDC_MBOX alone causes SPU driver only be
selected for platforms having PDC_MBOX driver, for example
spu2 driver is never enabled in stingray due to this.

As SPU driver is directly dependent on generic MAILBOX
framework API's, so we should be adding MAILBOX as
depends on.

Bug:SOC-5397
Change-Id: I6e8ea6624014e479f0e3490e70591c67d63d29a0
Signed-off-by: Raveendra Padasalagi <raveendra.padasalagi@broadcom.com>
Reviewed-on: http://gerrit-ccxsw.broadcom.net/43157
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Reviewed-by: CCXSW <ccxswbuild@broadcom.com>
Reviewed-by: Ray Jui <ray.jui@broadcom.com>
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Tested-by: vpx_smoketest status <vpx_smoketest@broadcom.com>
Tested-by: CCXSW <ccxswbuild@broadcom.com>
Reviewed-by: Scott Branden <scott.branden@broadcom.com>