VOIP-7049: Enable ALSA Loopback device
Enable ASLA's Loopback sound device.  The device is being enabled
as a module.  A static build of this driver will cause it to be
the first sound card instantiated, shifting the enumeration of
our existing sound cards by 1. Changing the enumeration will
be a problem for scripts and applications that have hardcoded
the cards value.  Creating the Loopback device as a module will
start the Loopback  device after our statically built sound card.

Bug: VOIP-7049
Change-Id: I6a4bcc8e4d0a43a65410eca330aedabee3a047a9
Reviewed-on: http://gerrit-ccxsw.broadcom.net/43923
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Tested-by: vpx_smoketest status <vpx_smoketest@broadcom.com>
Reviewed-by: Ray Jui <ray.jui@broadcom.com>
Tested-by: CCXSW <ccxswbuild@broadcom.com>
Reviewed-by: CCXSW <ccxswbuild@broadcom.com>
Reviewed-by: Scott Branden <scott.branden@broadcom.com>