SOC-5582: Sync-up stingray dtsi files with upstreamed version
We had done changes in upstreamed stingray dtsi files (based
on suggestion from maintainers) which are not present in
bcm/master kernel.

These changes are following:
1. Use "ranges" for each Stingray SOC block (SCR, HSLS, etc)
and use relative addresses in "reg" property of device DT nodes
2. Keep the "chosen" DT node to be minimal because most of the
information in "chosen" DT node are provided by bootloader
3. Don't specify GIC_CPU_MASK for PPIs

This patch updates all stingray-xxx.dtsi files to be in
sync with upstreamed version.

Signed-off-by: Anup Patel <anup.patel@broadcom.com>
Change-Id: I88ad159fd035ce26857478db95e2d316a65ee9d6
Reviewed-on: http://gerrit-ccxsw.broadcom.net/44277
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Reviewed-by: CCXSW <ccxswbuild@broadcom.com>
Tested-by: CCXSW <ccxswbuild@broadcom.com>
Reviewed-by: Scott Branden <scott.branden@broadcom.com>