SOC-5582: Allow No-IOMMU mode for IOMMUs with bypass capability
vfio: Allow No-IOMMU mode for IOMMUs with bypass capability

Not allowing No-IOMMU mode for devices already having
iommu_ops on their bus is little conservative.

We now have IOMMU (such as ARM SMMU) which can bypass
transcations for which IOMMU domain is not configured
hence No-IOMMU mode should not be allowed when iommu_ops
are available and IOMMU_CAP_BYPASS is not available.

Signed-off-by: Anup Patel <anup.patel@broadcom.com>
Change-Id: If523f9fbceda32792cf3bedc9642de187c96a101
Reviewed-on: http://gerrit-ccxsw.broadcom.net/44448
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Reviewed-by: CCXSW <ccxswbuild@broadcom.com>
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Tested-by: CCXSW <ccxswbuild@broadcom.com>
Reviewed-by: Scott Branden <scott.branden@broadcom.com>