VOIP-7077: clk: iproc: Remove unused code
The flag IPROC_CLK_SET_RATE_PARENT is an artifact from a feature that
was added for the audio clock tree.  This feature has subsequently been
removed so this flag is no longer needed.  Also, remove a function
prototype for which there is no function implemented.
Removing this code will also keep our branch in sync the upstream version
of the code.

Change-Id: I6a50241ac63a568a847d8cbf04a82402f84de37e
Reviewed-on: http://gerrit-ccxsw.broadcom.net/44553
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Reviewed-by: Ray Jui <ray.jui@broadcom.com>