SOC-5780: mdio_bus: Add suspend_late and early_resume callbacks
suspend_late and early_resume callbacks are required for mdio bus
when PHY(s) on mdio bus need to be suspended 'after' corresponding
controller has been suspended and need to be powered on 'before'
the controller resume sequence starts. This is the case with Broadcom
USB 3.0 PHY registered under mdio bus for Northstar2 platform.

Change-Id: I0a9d1158995d8d0d45a40bf487fa7c29ce373875
Signed-off-by: Abhishek Shah <abhishek.shah@broadcom.com>
Reviewed-on: http://gerrit-ccxsw.broadcom.net/46103
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Reviewed-by: CCXSW <ccxswbuild@broadcom.com>
Reviewed-by: Ray Jui <ray.jui@broadcom.com>
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Tested-by: vpx_smoketest status <vpx_smoketest@broadcom.com>
Tested-by: CCXSW <ccxswbuild@broadcom.com>
Reviewed-by: Scott Branden <scott.branden@broadcom.com>