SOC-5548: bpf: Add cross compiler support to xdp samples
The bpf makefile didn't support cross compiling which made it unusable
except on the build host. HOSTCC is set to CROSS_COMPILE if defined.
Additional HOSTCFLAGS and HOSTLDFLAGS can now be passed in through
EXTRA_HOSTCFLAGS and EXTRA_HOSTLDFLAGS.

The map_perf_test program doesn't compile on arm64. A fix is required to
use the correct syscall define.

An install target was added to install programs and the object files
they load at runtime.

Change-Id: I9067b255175990712e94ec96a3263b3681110dde
Reviewed-on: http://gerrit-ccxsw.broadcom.net/47103
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Reviewed-by: Jonathan Richardson <jonathan.richardson@broadcom.com>
Reviewed-by: Andy Gospodarek <gospo@broadcom.com>
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Tested-by: Jonathan Richardson <jonathan.richardson@broadcom.com>
Reviewed-by: JD Zheng <jiandong.zheng@broadcom.com>
Reviewed-by: Ray Jui <ray.jui@broadcom.com>
Reviewed-by: Scott Branden <scott.branden@broadcom.com>