VOIP-7158 Change SPI LCD to polling mode
Change SPI LCD on bcm958305_aud_base and bcm958305_ak4458 boards to
default polling mode.
When pl022 SPI is configured in interrupt mode it blocks interrupts
for the duration of the entire transfer causing other interrupts
to not be serviced in time.
DMA mode would be preferable, but currently not working due to a
RX FIFO overrun issue.

Change-Id: Ia5807794c37d48021a66865c2da2ff67cc3d2ce5
Reviewed-on: http://gerrit-ccxsw.broadcom.net/47370
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Reviewed-by: Ray Jui <ray.jui@broadcom.com>
Tested-by: Corneliu Doban <corneliu.doban@broadcom.com>
Reviewed-by: CCXSW <ccxswbuild@broadcom.com>
Tested-by: CCXSW <ccxswbuild@broadcom.com>
Reviewed-by: Scott Branden <scott.branden@broadcom.com>