SOC-6923 samples/bpf: Fix pt_regs issues when cross-compiling
BPF samples fail to build when cross-compiling for ARM64 because of incorrect
pt_regs param selection. This is because clang defines __x86_64__ and
bpf_headers thinks we're building for x86. Since clang is building for the BPF
target, it shouldn't make assumptions about what target the BPF program is
going to run on. To fix this, lets pass ARCH so the header knows which target
the BPF program is being compiled for and can use the correct pt_regs code.

Change-Id: I33041289e5572602cf02c6ea815f0100b091bb54
Acked-by: Alexei Starovoitov <ast@kernel.org>
Signed-off-by: Joel Fernandes <joelaf@google.com>
Acked-by: Daniel Borkmann <daniel@iogearbox.net>
Signed-off-by: David S. Miller <davem@davemloft.net>
Reviewed-on: http://gerrit-ccxsw.broadcom.net/57405
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Reviewed-by: Scott Branden <scott.branden@broadcom.com>