VOIP-7106: ASoC: cygnus: Use wrapper for io access
Use a wrapper function around the io accesses.  The plan is to
eventually use regmap instead of an io handle.  The regmap api
follows a different form than the readl/writel api.  This new
wrapper is the same as the regmap api, so when we eventually
move to regmap there will be little the change in this file.

Change-Id: Ib7473fc68314a38bd6a9abdb6a318ec1f94eba6d
Reviewed-on: http://gerrit-ccxsw.broadcom.net/47950
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Reviewed-by: CCXSW <ccxswbuild@broadcom.com>
Tested-by: CCXSW <ccxswbuild@broadcom.com>
Reviewed-by: Ray Jui <ray.jui@broadcom.com>
Reviewed-by: Scott Branden <scott.branden@broadcom.com>