VOIP-7257 brcmnand:iproc_nand disable onfi_set/get_feature commands
Disable the onfi set/get feature commands for now due to a problem
with some Micron nand devices that have on-die ECC.
The Micron specific initialization is determining if the on-die ECC
is supported (even if on-die ECC is not configured in device tree) by
enabling it using onfi_set_feature and checking using onfi_get_feature
then disabling it if the check found it enabled.
On cygnus this check finds that on-die ECC is not enabled when
apparently it is, resulting in leaving it eanbled and cause nand
corruption.

Change-Id: I53c363c520409b7608fcbdb53e4245c742c9a690
Reviewed-on: http://gerrit-ccxsw.broadcom.net/49292
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Reviewed-by: Ray Jui <ray.jui@broadcom.com>
Tested-by: Corneliu Doban <corneliu.doban@broadcom.com>
Reviewed-by: CCXSW <ccxswbuild@broadcom.com>
Tested-by: CCXSW <ccxswbuild@broadcom.com>
Reviewed-by: Scott Branden <scott.branden@broadcom.com>